move more stream handling into lib
